+++ CHANGELOG.md
@@ -23,6 +23,8 @@
 
 - `[blocksync]` validate blocksync response sender and signature count
   ([\#5860](https://github.com/cometbft/cometbft/pull/5860))
+- `[autofile]` skip fsync in `FlushAndSync` when no new data was written
+  ([\#5866](https://github.com/cometbft/cometbft/pull/5866))
 
 ### FEATURES
 
@@ -118,6 +120,7 @@
   ([\#5717](https://github.com/cometbft/cometbft/pull/5717))
 
 ### IMPROVEMENTS
+
 - `[consensus]` perf(consensus): skip fsync for unsigned internal messages (block parts) ([\#5695](https://github.com/cometbft/cometbft/pull/5695))
 - `[ci]`: add lp2p testnet ([\#5643](https://github.com/cometbft/cometbft/pull/5643))
 - `[mempool]` feat!(p2p): introduce follower-mode. Improve lib-p2p integraap access